pyproject.toml
[project]
name = "ragcoach"
version = "0.1.0"
description = "RAGCoach"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.2",
    "pydantic-settings>=2.6.1",
    "qdrant-client>=1.11.0",
    "sentence-transformers>=3.3.1",
    "pdfplumber>=0.11.4",
    "ollama>=0.11.6",
    "fastapi>=0.115.5",
    "uvicorn>=0.32.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
    "ruff>=0.8.2",
    "black>=24.10.0"
]
